Fixed explanation exc1, exc2 and solved exc3

+++ src/com/laingard/Excercise2.java
@@ -2,16 +2,16 @@ package com.laingard;
 
 import java.util.Scanner;
 
+/* Exercise 2
+ *
+ * Para este ejercicio tendréis que crear una función que reciba un precio y devuelva el precio con el IVA incluido.
+ * */
+
 public class Excercise2 {
     public static void main(String[] args) {
-
-    /* Exercise 2
-     *
-     * Para este ejercicio tendréis que crear una función que reciba un precio y devuelva el precio con el IVA incluido.
-     * */
     Scanner scanner = new Scanner(System.in);
         System.out.println("Introduce el precio: ");
-    double price = scanner.nextInt();
+    double price = scanner.nextDouble();
         System.out.println(priceWithIVA(price));
 }
     static double priceWithIVA(double price){
